validate expected git branch; fail if no match
it can be painful to run batch jobs against multiple repos is one or more of those repos is not on the expected branch.

this adds a mechanism to specify a branch (defaults to master), and batch scripts will fail if ALL of the expected git repos are not on that branch

also pass $BRANCH environment variable down to MBML as $repo:branch so we can use it for documentation generation, etc.